LayoutTests/imported/w3c/web-platform-tests/event-timing/modal-dialog-interrupt-paint-expected.txt

ALERT:
Click me.

FAIL Event Timing: showing modal dialogs is an alternative end point. assert_less_than: Event duration measurement should stop at the modal dialog showing time, which should be before processingEnd. expected a number less than 1176 but got 1182
